requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
lxml==5.3.0
multidict==6.1.0
openai==1.51.2
pdfminer.six==20240706
plum-dispatch==1.7.4
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-docx==1.1.2
requests==2.32.3
sniffio==1.3.1
Spire.Doc==12.7.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.14.0
